requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Audio Processing
openai-whisper==20231117
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.4

# Text-to-Speech
elevenlabs==0.2.27
gTTS==2.4.0
pygame==2.5.2

# AI/LLM
anthropic==0.7.8
ollama==0.1.7
openai==1.3.8

# Vector Database & RAG
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.4

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Smart Home Integration
homeassistant-api==0.1.10
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
loguru==0.7.2
PyYAML==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1